build_lm32.mk: fix more symbol assignments (but not all of them)
This applies the same idea explained in the previous commit the .data
and .rodata.  The commit has no effect because nobody is using _fdata
and _frodata, which were incorrectly defined (they both pointed to the
end of the section, not to the beginning as expected).

This does not fix _fbuildid, _fshared and _ftext because the change
needs to be made differently. Again, nobody is using those so the bug
has no effect.