Extended the temperature sensor reading.
Now temperature value in degree can be read at the <base>+4 address as
32-bit signed integer. The default sensor data reading at the <base>
address is still supported, but consider the MSB as a validity flag.
If ADC is failed or timed out, then the wb slave device for the temperature
sensor will return sensor data with the MSB bit set and temperature value
in degree higher than 255.