Makefile: use ':' and '::' consistently
A double-colon rule with no prerequisites is like a phony
target, which is what we want.

A double-colon rule with prerequisite is a different thing
(a trange one), and is not what we want. So instead of using
some of one kind and some of another, aways use : if there are
prerequisites (which in this case all phony targets).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>