syn/build.mk: do not include $(TARGET).dep
This $(TARGET).dep is generated by the rule for .sof files.
It is then optionally included by Makefile itself.

Unfortunately, this huge single-line file says "-library work",
which means our make will try to make ./-library and has no rule
for it.

By being included optionally, it is not included at the first run
of the build, so no error happens on first run or after a make clean.

Considering it is never really used, we won't include it. It is
unkown, so far, whether it is used by quartus or not. So let's leave
it in place.

Bug from 2013, when build.mk was initially written.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>