Makefile: do not reconfig wrpc-sw if already configured
If we reconfigure, we force a rebuild as well. Moreover, if/when
somebody is working on wrpc-sw their config mods would be lost.

This commit has no effect on the user who is not working on wrpc-sw,
but those who do must be sure the build gw with the proper sw config.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>