lm32: handles message-signaled interrupts (MSIs) sent by ECA
The ECA unit is capable to send MSI on certain conditions, such as
producing valid actions in its action channel on reception of timing
messages.

ECA evaluates incoming timing messages against pre-configured action rules
in its condition table. If event identifier match between the timing
message and existing action rules is detected, then ECA produces actions at
the designated time, outputs them to dedicated action channel and sends
MSI interrupt to hooked interrupt handler.

In order to enable MSI generation both the condition table and MSI registers
in action channel must be properly configured:
- conditions (action rules) to produce actions in dedicated action channel
- destination path of MSI handler

Once ECA sends MSI to a handler specified in the destination path it
disables the generation of MSIs until pending action in the action channel
is handled.
Every time an action is output, it is held in a queue connected to
an action channel for further handling. When the action is picked up and
desired task is done by a handler, then it must be popped from the queue,
otherwise the queue will overflow. With the popping the valid flag of that
action is reset. However that's still not sufficient to enable interrupt
generation.
ECA action channels have a set of condition counters for its actions:
valid, overflow and failed. Eventually these counters must be read and
cleared to enable the generation of MSIs.

Therefore, the handling of ECA MSIs requires two steps:
- read and clear action channel counters
- pop pending actions from a queue connected to that action channel