syn/build.mk: sort qsf file to minimize differences
The qsf file is a list of global assignments, and every line
is independent. Unfortunately, the content is sometimes re-saved
by quartus in a completely different order, but the content is
equivalent.

For example, commit cad275d reports 719 changes in scu_control.qsf,
wherease it only grew by 3 new lines.

After this commit, and respective builds, we'll have a single big
diff for each qsf file, but later on the changes will be minimal and
user-readable.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>