lm32 linker script (and *shared_mmap.h): customize in a different way
The previous build.mk with build_lm32.mk was creating the
linker script out of make variable, using "echo $(LDS) > ram.ld"
This makes the linker script itself unreadable, as the
real contents is in "LDS +=" lines.

This commit, instead, changes the ram size and other parameters
from a template linker script, based on the wrpc-sw one (which
already has a little more features). The same applies to shmem.h
headers, because they export the same numbers to user space.

The template linker script is called ram.ld.S, in include-lm32/
(sane applies to shared_mmap.h.S). Adding a project override
later will be easy ("RAM_LD_S ?= $(INCPATH)/ram.ld.S").

As a side effect, configurable values are now called CONFIG_*.
The ones defined in project-specific files are unchanged,
to avoid interfering with all the projects and be able
to cherry-pick this to master or otherwise.

This is a list of all affected make variables:

 * RAM_OFFS ->  CONFIG_RAMADDR
 * BOOTL_SIZE -> CONFIG_BOOTSIZE
 * BUILDID_SIZE -> CONFIG_BUILDIDSIZE

   all 3 are constants in build_lm32.mk
   substituted in .lds, shmem.h, buildid.c

 * INTADR_OFFS -> removed
   (it was a duplicate of RAM_OFFS)

 * RAM_SIZE -> CONFIG_RAMSIZE
   still provided by each project as "RAM_SIZE"
   substituted in .lds, shmem.h, buildid.c

 * SHARED -> CONFIG_SHAREDSIZE
   still provided by each project as "SHARED_SIZE"

 * UCTARGET -> CONFIG_TARGET
   used locally, only for the ifndef of the header

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>