ftmfw: call pp_printf at least once
scu2/3 calls both pp_printf and mprintf, so we chose to remove
mprintf, providing it as an alias.

Unfortunately, ftmfw is never calling pp_printf, so the function is
discarded by --gc-sections; when the linker finally uses
PROVIDE(this,that) we have mprintf=0, and the link finally fails.

So let's call pp_printf at least once, to avoid discarding it. The
rest follows at no cost.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>