syn/build: add assert (panic) and stack-check
assert calls panic if the condition is false. Panic prints the
message every second. The code comes from wrpc-sw, where I used it
(so it's tested).

If the specific project wants to disable panic, it must say

   CONFIG_NO_ASSERT = y

in Makefile before including build.mk. In that case, no panic
happens and assertions are ignored.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>