eb_fifo: alternative implementation running in ghdl
This is a demo of how to run a ghdl testbench using sources from bel_projects.
It needs the arria5 vendor libs complied with ghdl.
The location of the vendor libs needs to be put in the first line of makefile.
There is also an alternative implementation of eb_fifo, called my_eb_fifo, with identical behavior.