lm32: fixed an issue getting zero value when reading the valid counter
Each ECA channel has a valid counter register that keeps the number of
occurrences of a valid action. Once the register is read, then its value
is cleared. The user can read the register by polling it or on reception
of MSI with the valid action flag.
The issue was that sometimes zero value was read from the valid counter
register after receiving an MSI with the valid action flag. It was caused
by the random selection of the sub_channel.