wb_adapter_simulation: improved testbench.
Slave can be confiured (via comments) to support different behavior.
With the current modification of wb_slave_adapter.vhd, The slave works
correctly if the ACK is one clock cycle long. The ACK may be CYC and
STB or a response to the rising edge of STB delayed by an arbitrary
number of clock cycles. Main point is that it is never longer then one
clock cycle.