fbas: use old SCP protocol for back-compatibility
scp was failed on tsl101 by returning an error:
sh: /usr/libexec/sftp-server: not found

Cause: tsl101 (Rocky9) uses the newest OpenSSH, which uses SFTP server for
scp. But the SFTP server is not installed on tsl1010.
With -O option, the old SCP protocol is used instead of SFTP server.