fbas: update 'restart_test_rx_rate'
The number of the DM broadcast and received messages are compared directly.

Results of the messaging delay measurement is unavailable: missing a proper
sender ID (0xffffff is used instead of DM MAC) in the parameter field of
the DM schedule (my_mps_rx_rate_1.dot).

'run_pattern' returns the number of sent messages in hex, therefore, the
recipe line ends with 'echo' to insert the new line behind this result.